refactor: unify LanguagePicker component
