+++ src/components/ui/language-picker.tsx
@@ -1,0 +1,19 @@
+import { useLocale } from "@/lib/i18n/hook";
+import { Locale, LOCALE_LABELS, LOCALES } from "@/lib/i18n/types";
+
+export function LanguagePickerSection() {
+  const { locale, setLocale } = useLocale();
+  return (
+    <select
+      value={locale}
+      onChange={(e) => setLocale(e.target.value as Locale)}
+      className="w-full px-3 py-2 rounded-lg border-2 border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-800 focus:ring-2 focus:ring-blue-500 focus:border-blue-500 text-sm"
+    >
+      {LOCALES.map((code) => (
+        <option key={code} value={code}>
+          {LOCALE_LABELS[code]}
+        </option>
+      ))}
+    </select>
+  );
+}
